+++ ec/src/scalar_mul/mod.rs
@@ -4,8 +4,10 @@ pub mod wnaf;
 pub mod fixed_base;
 pub mod variable_base;
 
-use crate::PrimeGroup;
-use crate::short_weierstrass::{Affine, Projective, SWCurveConfig};
+use crate::{
+    short_weierstrass::{Affine, Projective, SWCurveConfig},
+    PrimeGroup,
+};
 use ark_ff::{AdditiveGroup, Zero};
 use ark_std::{
     ops::{Add, AddAssign, Mul, Neg, Sub, SubAssign},