+++ CHANGELOG.md
@@ -2,13 +2,14 @@
 
 ## Pending
 
+- [\#659](https://github.com/arkworks-rs/algebra/pull/659) (`ark-ec`) Add Elligator2 hash-to-curve map.
+
 ### Breaking changes
 
 - [\#577](https://github.com/arkworks-rs/algebra/pull/577) (`ark-ff`, `ark-ec`) Add `AdditiveGroup`, a trait for additive groups (equipped with scalar field).
 - [\#593](https://github.com/arkworks-rs/algebra/pull/593) (`ark-ec`) Change `AffineRepr::xy()` to return owned values.
 
 ### Features
-- [\#xxx](https://github.com/arkworks-rs/algebra/pull/xxx) (`ark-ec`) Add Elligator2 hash-to-curve map.
 
 ### Improvements